Consolidate IO module under src/hydro/io and include/hydroc/io
Moves all HDF5 and simulation export code into a dedicated IO module:

- Added new public headers: include/hydroc/io/{h5_reader,h5_writer,simulation_export}.h
- Moved implementations to src/hydro/io/
- Updated all includes and CMake entries
- Removed obsolete top-level redirect headers (h5fileinfo.h, h5_writer.h, simulation_exporter.h)

Public API and behaviour unchanged. Repository IO structure is now clean, consistent, and easier to navigate.
