+++ src/hydro/waves/irregular_wave.h
@@ -9,7 +9,7 @@
 #include "wave_base.h"
 #include "wave_utilities.h"
 
-#include <hydroc/h5fileinfo.h>
+#include <hydroc/io/h5_reader.h>
 
 #include <string>
 #include <vector>
